Move polling hook to its own file
