+++ src/user/Notifications.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from "react"
+import React, { useRef, useState } from "react"
 import { Popover } from "@headlessui/react"
 import { useSelector } from "react-redux"
 
@@ -7,8 +7,8 @@ import * as userNotification from "./notification"
 import * as user from "./userState"
 import { useTranslation } from "react-i18next"
 import * as appState from "../app/appState"
-import * as request from "../request"
 import broadcastIcon from "./broadcast.svg"
+import { useNotificationLongPolling } from "./useNotificationLongPolling"
 
 interface Message {
     text: string
@@ -98,97 +98,6 @@ export const NotificationPopup = () => {
             }}>X</span>
         </div>
     </div>
-}
-
-/** Automatically fetch notifications every X minutes when logged in */
-const useNotificationLongPolling = () => {
-    // Polling interval sequence, stopping after the final interval
-    const NOTIFICATION_POLL_INTERVALS_MS = [5, 5, 5, 30, 60, 60].map(min => min * 60 * 1000)
-
-    const isLoggedIn = useSelector(user.selectLoggedIn)
-
-    useEffect(() => {
-        if (!isLoggedIn) return
-
-        let timeoutId: number | null = null
-        let intervalIndex = 0
-        let isPolling = false
-        let hasFetchedInitialNotifications = false
-        let latestNotificationCount: number | null = null
-
-        const fetchNotifications = async () => {
-            try {
-                const result = await request.getAuth("/users/notifications")
-                if (Array.isArray(result)) {
-                    // Reset to the shortest polling interval if we detect new notifications
-                    if (latestNotificationCount !== null && result.length > latestNotificationCount) {
-                        intervalIndex = 0
-                    }
-                    latestNotificationCount = result.length
-                    userNotification.loadHistory(result)
-                }
-            } catch (error) {
-                console.error("Error fetching notifications:", error)
-            }
-        }
-
-        const clearScheduledPoll = () => {
-            if (timeoutId == null) return
-            window.clearTimeout(timeoutId)
-            timeoutId = null
-        }
-
-        const scheduleNextPoll = () => {
-            if (!isPolling || timeoutId != null) return
-            if (intervalIndex >= NOTIFICATION_POLL_INTERVALS_MS.length) {
-                isPolling = false
-                return
-            }
-
-            const interval = NOTIFICATION_POLL_INTERVALS_MS[intervalIndex]
-            timeoutId = window.setTimeout(async () => {
-                timeoutId = null
-                intervalIndex += 1
-                await fetchNotifications()
-                scheduleNextPoll()
-            }, interval)
-        }
-
-        // Immediately fetch on page load, then schedule the next poll
-        const startPolling = () => {
-            if (isPolling) return
-            isPolling = true
-            if (!hasFetchedInitialNotifications) {
-                hasFetchedInitialNotifications = true
-                fetchNotifications()
-            }
-            scheduleNextPoll()
-        }
-
-        const stopPolling = () => {
-            isPolling = false
-            clearScheduledPoll()
-        }
-
-        // Only poll when the page is visible to avoid unnecessary requests
-        const onVisibilityChange = () => {
-            if (document.visibilityState === "visible") {
-                startPolling()
-            } else {
-                stopPolling()
-            }
-        }
-
-        // start based on initial visibility
-        onVisibilityChange()
-
-        document.addEventListener("visibilitychange", onVisibilityChange)
-
-        return () => {
-            stopPolling()
-            document.removeEventListener("visibilitychange", onVisibilityChange)
-        }
-    }, [isLoggedIn])
 }
 
 /** Small blue bullhorn indicator used for broadcast announcements. */
